Fixed issue controllers

+++ Controllers/IssueControllers/IssueItemController.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -7,6 +8,7 @@ using Microsoft.Extensions.Logging;
 using TF47_Backend.Database;
 using TF47_Backend.Database.Models.Services;
 using TF47_Backend.Dto.RequestModels;
+using TF47_Backend.Dto.ResponseModels;
 using TF47_Backend.Services;
 
 namespace TF47_Backend.Controllers.IssueControllers
@@ -16,12 +18,12 @@ namespace TF47_Backend.Controllers.IssueControllers
     [ApiController]
     public class IssueItemController : ControllerBase
     {
-        private readonly ILogger<Issue> _logger;
+        private readonly ILogger<IssueItemController> _logger;
         private readonly DatabaseContext _database;
         private readonly IUserProviderService _userProviderService;
 
         public IssueItemController(
-            ILogger<Issue> logger,
+            ILogger<IssueItemController> logger,
             DatabaseContext database,
             IUserProviderService userProviderService)
         {
@@ -32,6 +34,7 @@ namespace TF47_Backend.Controllers.IssueControllers
         }
 
         [HttpPost("")]
+        [ProducesResponseType(typeof(IssueItemResponse), 200)]
         public async Task<IActionResult> CreateIssueItem([FromBody] CreateIssueItemRequest request)
         {
             var issueGroup =  await _database.Issues.FirstOrDefaultAsync(x => x.IssueId == request.IssueId);
@@ -52,18 +55,23 @@ namespace TF47_Backend.Controllers.IssueControllers
             await _database.SaveChangesAsync();
 
             return CreatedAtAction(nameof(GetIssueItem), new {issueItemId = issueItem.IssueId},
-                issueItem);
+                new IssueItemResponse(issueItem.IssueId, issueItem.Author.UserId, issueItem.Author.Username,
+                    issueItem.Message, issueItem.TimeCreated, issueItem.TimeLastEdited));
         }
 
         [HttpGet("{issueItemId:int}")]
+        [ProducesResponseType(typeof(IssueItemResponse), 200)]
         public async Task<IActionResult> GetIssueItem(int issueItemId)
         {
             var issueItem = await _database.IssueItems
+                .Select(x => new IssueItemResponse(x.IssueItemId, x.Author.UserId, x.Author.Username, x.Message,
+                    x.TimeCreated, x.TimeLastEdited))
                 .FirstOrDefaultAsync(x => x.IssueItemId == issueItemId);
             return Ok(issueItem);
         }
 
         [HttpPut("{issueItemId:int}")]
+        [ProducesResponseType(typeof(IssueItemResponse), 200)]
         public async Task<IActionResult> UpdateIssueItem(int issueItemId, [FromBody] UpdateIssueItemRequest request)
         {
             var issueItem = await _database.IssueItems.FindAsync(issueItemId).AsTask();
@@ -81,15 +89,30 @@ namespace TF47_Backend.Controllers.IssueControllers
 
             await _database.SaveChangesAsync();
 
-            return Ok(issueItem);
+            return Ok(new IssueItemResponse(issueItem.IssueId, issueItem.Author.UserId, issueItem.Author.Username,
+                issueItem.Message, issueItem.TimeCreated, issueItem.TimeLastEdited));
         }
 
         [HttpDelete("{issueItemId:int}")]
         public async Task<IActionResult> DeleteIssueItem(int issueItemId)
         {
             var issueItem = await _database.IssueItems.FindAsync(issueItemId);
-            _database.Remove(issueItemId);
-            await _database.SaveChangesAsync();
+
+            if (issueItem == null) return BadRequest("IssueItem provided does not exist");
+            
+            try
+            {
+                _database.Remove(issueItemId);
+                await _database.SaveChangesAsync();
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError("Failed to remove issueItem with id {id} from the database: {message}", issueItemId,
+                    ex.Message);
+                return Problem("Found issueItem in database but unable to remove it", null, 500,
+                    "Failed to remove IssueItem");
+            }
+
             return Ok();
         }
     }
